Refactor the codes into sub folders
