+++ .circleci/config.yml
@@ -21,4 +21,6 @@ jobs:
           name: run tests
           command: |
             . venv/bin/activate
-            pytest -v --cov=Group --cov=User --cov=DivvyNumericTester
+            coverage run -m pytest
+            coverage report -m
+            coverage html
